Stitching together a .White-flag
Funny that I have the elements but cannot connect the dots.  This is
why they say that calculus is based on basic math.  I’ve got to learn
how to think like a jr. dev.  That is what the first two weeks of utter
failure has taught me.. You were both right.